+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * The `--only` option now enables the given cop in case it is disabled in configuration. ([@jonas054][])
 * Fix path resolution so that the default exclusion of `vendor` directories works. ([@jonas054][])
 * [#908](https://github.com/bbatsov/rubocop/issues/908): Fixed hanging while auto correct for `SpaceAfterComma` and `SpaceInsideBrackets`. ([@hiroponz][])
+* [#919](https://github.com/bbatsov/rubocop/issues/919): Don't avoid auto-correction in HashSyntax when there is missing space around operator. ([@jonas054][])
 
 ## 0.19.1 (17/03/2014)